Clear all stations

+++ service/todo.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
+	"strings"
 
 	"github.com/TechBowl-japan/go-stations/model"
 )
@@ -115,7 +116,33 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 
 // DeleteTODO deletes TODOs on DB by ids.
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
+	if len(ids) == 0 {
+		return nil
+	}
 	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
+	query := fmt.Sprintf(deleteFmt, strings.Repeat(",?", len(ids)-1))
+	args := make([]interface{}, len(ids))
+	for i, id := range ids {
+		args[i] = id
+	}
+	result, err := s.db.ExecContext(ctx, query, args...)
+	if err != nil {
+		return err
+	}
+	rows, err := result.RowsAffected()
+	if err != nil {
+		return err
+	}
+	if rows == 0 {
+		return model.ErrNotFound{Message: "ID not found"}
+	}
+	// for i, id := range ids {
+	// 	fmt.Printf("%d: %s\n", i, id)
+	// 	dmt := fmt.Sprintf(`DELETE FROM todos WHERE id IN (?%s)`, id)
+	// }
+	// if err != nil {
+	// 	return err
+	// }
 
 	return nil
 }
